control: cached loading for tls keys/certs from disk
Previously only the keys were loaded from disk, the chains were fetched
from the DB. This becomes a problem with certificate renewal (especially
when overwriting the key/cert files) because the new key is loaded from the disk
but no corresponding certificate is loaded into the DB
Additionally, loading certs from disk for each individual TLS session
establishment is not ideal, so we should have a cache.

Removed the separate X509KeyPairLoader and replaced it with
trust.SignerGen (extended to allow filtering certificates for key
usage). For this SignerGen, a CachingSignerGen implementation already
exists.

Note: The filtering of the certificates for the ext key usage now is
implemented in the SignerGen, based on a member field KeyUsage.
Alternatively, the filtering could have been implemented in the
DB/Loader level, as part of a ChainQuery. This approach was not taken
because it would seemingly lead to ambigous behaviour in the
CryptoLoader, where it would not have been clear whether only matching
certificates are inserted to the DB.
Yet another option would have been to expose the KeyUsage option to the
SignerGen interface (something like `signerGen.Generate(ctx, WithExtKeyUsage(ExtKeyUsageServerAuth))`,
but this would make the CachingSignerGen much more complicated.