build: update gobra to support the new rules of gofmt (#4279)
* Update Gobra to latest version to be compatible with the new gofmt rules

Update to latest version and adapt changed config options.

Introduce a new file (util_spec.gobra) because Gobra now requires that all imported packages are populated (i.e., have at least one annotated file). This is a recent change which has to do with the support for specifying and verifying initialization code. Maybe this check is a bit too strict, so it may be dropped in the future. At any rate, util_spec.gobra will be useful to verify the method String() in infofield.go, when Gobra supports the fmt package.