slayers: combine AddrType/Len (#4160)
AddrType/Len fields are always set and read together. The AddrType field
is meaningless without the Length.

This change combines these types/fields into one value that uniquely
identifies the address type, simplifying the processing of this address
length/type.
The length (in bytes) of the address type can be obtained with a method
call. There is (intentionally) no conversion functionality to extract
only the type (as this is meaningless on its own), or to construct a
AddrType from type/length parts.

This change prepares for possible further refactoring of the interface
for setting/getting src/dst addresses on slayers.SCION.