braccept: fix swapped AS in diagram and test cases
The diagram for the test topology had swapped the relations to neighbor
ASes ff00:0:2 (is peer, was shown as parent) and ff00:0:3 (is parent,
was shown as peer).

Various test cases used the link to ff00:0:2 (interface number 121),
where evidently a parent link was supposed to be used.
Replaced these occurrences with the link to ff00:0:3.

Fixed the link relations in the diagram.
At the same time, fixed some smaller issues in the diagram (wrong label
for sibling router C, wrong label for interface 211, typo in IP for
interface 511).
Removed the embedded draw.io "pages" for other test topology setups that
no longer exist.
Renamed to *.drawio.png, as a hint that this contains an embedded
draw.io scene.