router: add checks for intra-segment forwarding (#4294)
Similar to the existing checks for segment switching, add checks for the
case in which the router forwards a packet along a single segment. This
check only allows child-to-parent, parent-to-child, and core-to-core
combinations, as well as packets coming from an internal interface
(peering links are not yet considered).
This explicit check ensures valley-free forwarding, even if an attacker
knows all AS hop keys.
Invalid hops result in an SCMP Error message with type `Parameter
Problem` and code `Invalid Path`.

Add new braccept cases for the invalid hops child-to-child and
parent-to-parent. Additional cases would be child-to-core and
core-to-child, but these cannot be checked in the current braccept
framework.
Adapt existing braccept case SCMPTracerouteEgress, where a child-to-peer
hop was used, which is currently not considered valid (peering links are
not supported). It appears that this test case did not specifically
intend to use a peering link (as the path only consists of a single
segment).

Aside: restructure parameters for packSCMP in router to make calls more
compact and increase readability.