slayers: type-safe SetNetworkLayerForChecksum
Change parameter for slayers.UDP/SCMP.SetNetworkLayerForChecksum to
*slayers.SCION, from previously gopacket.Layer. Remove error from
signature, as explicit type check is no longer needed.

Note that in the gopacket/layer.UDP and TCP layers,
SetNetworkLayerForChecksum takes a generic gopacket.Layer for because it
accepts either IPv4 or IPv6.
There is no reason to keep this for the slayers layers, there is no
common interface covering this functionality.

Cleanup exceptions for errcheck in golangci-lint: removed all files that
use SetNetworkLayerForChecksum from exception list, assuming that not
checking the return code of this function was the errcheck violation.
Fix various small other errcheck issues; in particular, one frequent
case was to explicitly ignore the error from calling
gopacket/layers.UDP.SetNetworkLayerForChecksum in the braccept cases.