db: check tx.Rollback error / serrors: add Join
Check returned error of tx.Rollback where this had previously just been
a `defer tx.Rollback`. The errcheck linter (rightly) complains about this.
The Rollback was previously also be called after a successful commit, in
which case it would have returned the error `ErrTxDone`. Now, Rollback
is only called in error cases.

Not fixed the `defer tx.Rollback` pattern in colibri code, as this
appears to be pretty much dead code by now.

Add a `serrors.Join` helper function to make it easier to return
multiple potential errors. This API is modeled closely after
`errors.Join`, which will be available in the go standard library in the
next release (go 1.20).