github: add codeowners (#4306)
Add CODEOWNERS file. For details about the file format, see
https://help.github.com/articles/about-codeowners/

The overall owner over the implementation is the SCION Association's
Technical Committee Implementation, represented as @scionproto/scion-core-team.
Note that the CODEOWNERS file does not contain a '*' pattern to specify
this, as this would notify the entire core team by default. Instead, we
only define owners for specific modules that require particular
attention.