slayers: reset Options in extension DecodeFromBytes (#4308)
Reset the extension's Options field during DecodeFromBytes to avoid
including old values when reusing the extension object.
Add tests for this behavior with  TestEndToEndExtnDecodeReuse and
TestHopByHopExtnDecodeReuse.

Previously, if we were to decode an HopByHopExtn into a receiver e, then
the field e.Options would never be erased and this slice would accumulate all
the Options parsed in previous calls to DecodeFromBytes.

Furthermore, fix a type assertion bug (also related to extension
headers) in the router, which fortunately affects only a diagnostics message;
In the function nextHdr in the router, EndToEndExtn and HopByHopExtn
types can never occur, as only the EndToEndExtnSkipper and
HopByHopExtnSkipper types are actually used. Fixed this for now, but
this seems brittle and should be restructured in some form.