control: close connection in SVC resolver
When performing a service address resolution with svc.Resolver.LookupSVC,
a new connection was opened but never closed.
One possible reason for this bug was that the unclear responsibility for
closing the connection; the svc.RoundTripper.RoundTrip function closes
the connection if the context ends, but not otherwise.

Fixed always closing the conn after the request in LookupSVC. Also move
CloseConnOnDone here too, to make it clearer where the connection is
closed.