fixup: different approach, close in CloseConnOnDone
When performing a service address resolution with svc.Resolver.LookupSVC,
a new connection was opened but never closed.
One possible reason for this bug was that the unclear responsibility for
closing the connection; the svc.RoundTripper.RoundTrip function closes
the connection if the context ends, but not otherwise.

Fix this by always closing the conn in CloseConnOnDone.
The cancel function of CloseConnOnDone now also blocks until the
connection is closed (relevant mostly for testing) and returns any
errors from closing.
Also move the CloseConnOnDone to LookupSVC, so the responsibility for
closing the connection is right next to where it's opened.