control: cached loading of keys/certs from disk for TLS (#4281)
Previously only the keys were loaded from disk, the chains were fetched
from the DB. This becomes a problem with certificate renewal (especially
when overwriting the key/cert files) because the new key is loaded from the disk
but no corresponding certificate is loaded into the DB
Additionally, loading certs from disk for each individual TLS session
establishment is not ideal, so we should have a cache.

Removed the separate X509KeyPairLoader and replaced it with
trust.SignerGen (extended to allow filtering certificates for key
usage). For this SignerGen, a CachingSignerGen implementation already
exists. The new TLSCertificateLoader wraps this SignerGen and converts the
resulting Signer to a tls.Certificate.
Loading the certificates is no longer part of trust.TLSCryptoManager.
Thus, rename TLSCryptoManager to TLSCryptoVerifier.