seg: fix AddASEntry with multiple PeerEntries
During creating the signed protobuf body for an ASEntry in
PathSegment.AddASentry, a reference to a member of the loop variable
(`peer.HopField.MAC[:]`) was stored into newly created protobuf
PeerEntries. As a result, all protobuf PeerEntries ended up with the
same Mac buffer reference.
Fixed by the usual `peer := peer`.

Add tests for the case of multiple peers, both on the beacon extender
and the path segment level.