pkt/spao: support empty path value as input for SPAO MAC computation (#4317)
Implement support for empty path values as inputs for the SPAO MAC
computation.
This is currently not applicable when using the SPAO with DRKey as the
timestamp defined to be relative to the info field, which simply does
not exist for empty paths. This will be addressed by the changes
proposed in scionproto/scion#4300, enabling the use of SPAO also for AS
internal communication.