doc: fix python version in readthedocs config
In #4319, the python requirements file was generated for python3.8 but
our build configuration file still referred to 3.7, resulting in broken
builds. Also bump the ubuntu version to latest LTS, because, why not.

Also move the .readthedocs.yaml file to the project root, where this
should be located according to the documentation.
The readthedocs build log (confusingly) mentioned that this file was not
present. Either this was a spurious error message or the file was indeed
ignored and it just happened to match the defaults.