tools: add editor and lint configuration
Add a standardized vscode configuration that should lead to a more
consistent code base. Additionally, switch to markdownlint-cli that is
well supported in vs code and other editors.

The rules are kept the same, but we extend linting to our documentation
to. To satisfy the linter, we switch to myst parser for sphinx which
allows us to drop the hacky html tags in the markdown files.
Myst is as powerful as RST with a more sane syntax.