topology: deprecate unused attributes (#4333)
Deprecate unused attributes `authoritative`, `voting` and `issuing` from
the topology.json configuration. These are no longer considered by any
service or tool and the functionality has been moved into the cs.toml
(containing a `[ca]` section for "issuing" ASes) or into the TRC and the
configuration of the corresponding tooling. Only the `core` attribute
option remains. In the long run, this should either disappear entirely
from the topology configuration or be changed to a boolean flag.

In the internal `private/topology`, remove the unused `CA() bool` from
the interface, and simplify the representation of the state from a list
of attributes to a simpler `IsCore`.

The goal here is mainly to avoid confusion caused by the unused
attributes.