doc: reorganize and extend documentation (#4325)
The documentation has been a somewhat unsorted bag of documents. It was mostly focused on contributors to this project, but some sections did target users of the software, without a clear distinction.

This PR attempts to bring some more structure into this. There are now three separate sections:
   - Technology:
     Explanations and Specifications. Explain SCION on a conceptual level.
     Should introduce concepts that a user or dev needs to make sense of the manuals etc.
     Target audience: anyone (users, developers, outsiders)
   - Reference manuals:
     Target audience: users of this SCION implementation (operator of SCION infrastructure or hosts), users of any of the SCION APIs.
   - Developer section:
     Target audience: contributors to this SCION implementation
     
In particular, the developer section now "hides" the design documents away a bit and the intended workflow for these documents is clarified.

There is still a lot to do, in particular for overview and introductory material and the manuals. I've started to work on the manual for the control service, but I've decided to defer this for a separate PR so that we can finalize the organizational changes.
Note that the content of some design documents, in particular the one on path policies is now relegated to the less findable design document section. The relevant information will be included in the manuals again.