slayers: fix decoding and serializing wide DT/ST
Both decoding and serializing of the address type fields (DstAddrType
and SrcAddrType) was truncating the most significant bit.
This has not caused any issues so far as no address type with type
greater than 1 is defined and used.

This bug was introduced in #4160, by consistently using the wrong (!)
value 0x7, instead of 0xF, to mask the lowest four bits. Fixed the
same mistake in copies of the serialization logic.