build: generate docs with write_source_files (#4357)
Use write_source_files to generate the docs. The write_source_files
comes with automatic tests that the generated files are up to date.
We create a wrapper in tools/lint which checks that the generated
files are up to date if we run bazel test with --config=lint.

`make lint` checks that the `write_all` target is up to date with all
the additional targets to run.