build: Update dependencies install script for rhel/fedora.
Many packages that are explicitly installed in the debian case were
left implicit in the rhel case (in the assumption that they were
part of the Software Development package group). While at it, separated
the command from the list of packages as is doen for debian.

The pip3 package was used before there is a chance to install it. Since
pip3/deps isn't needed by the rest of the install scripts. Do it last.