python: update pyyaml to 6.0.1 (#4370)
Update pyyaml dependency from 6.0.0 -> 6.0.1, as amendment to #4367.

The actual issue that had caused our builds to fail suddenly, was not
related to the license_file warning, but an AttributeError  related to a
"cython_sources" option. It appears that pyyaml had an unconstrained
version of for Cython as a build system  dependency. When Cython 3 was
released, this started to fail.
Pyyaml 6.0.1 now has a constraint for Cython version < 3, which should
keep things working.

Somewhat worrying, and not addressed by this change, is the fact that
the build system dependencies do not seem to be pinned by our
requirements.txt file.