storage: add a new sqlite implementation that has no C dependency.
Added the modernc go sqlite implementation and made it the default. The reason is that the mattn implementation links statically to C code, which creates a dependency on the glibc version of the build host. That version can be incompatible with the one contained in the test containers, causing the integraton tests to fail. There are other ways to fix the problem in general:

* Full cross-build for the specific test distribution.
* Building the test containers to match the build host.
* Both (ie. build and test for different platforms, including the build host's)

Each of those have different merits and require a lot of work. In the meantime removing that dependency puts the issue out of the critical path and lowers complexity.

The mattn sqlite implementation is retained as an option. To make it easy to switch back-and-forth between implementations, all direct references to the sqlite implementation are confined to exactly one go file per implementation: private/storage/db/sqlite_<impl>.go. To switch between them, just edit the srcs attribute in private/storage/db/BUILD.bazel. To make this possible, I had to prevent gazelle from updating that target. If you know of a better way, chime in.