build: Improve sqlite implementation selection method.
The build option is now compatible with go build.

To build with mattn sqlite, pass the flag "-tag sqlite_mattn". To build with modernc sqlite, pass the flag "-tag sqlite_modernc".

When building via Bazel, that flag is injected via .blazerc with the line:

"common --show_timestamps --define gotags=sqlite_modernc"

Gazelle also needs to be supplied with that tag. For gazelle gotags must be passed as:

"-build_tags tag1,tag2,..."

This is accomplished by adding a pair of config_setting() calls that translate the gotags define into a setting tag and by using a select keyed on that to pass the appropriate flag to Gazelle. Gazelle update-repos does not need to know about the go tag.

On the Makefile side, Gazelle is now called with "-args" to pass those arguments that aare variables in the Makefile, allowing the others to be chosen and appended by Blaze rules.

It was noticed that make licenses implies a full build and therefore should be done last. (At least not before gazelle.)