hidden_paths: Simplifications - the Resolver address now always has a dynamic port.
Doing this allows the CS's SCION address to have the same host/port as the CS's public address (by default) instead of being dynamic. This in turn makes it discoverable, which allows the hidden path registry to be reachable.

This approach allows the topology file to declare the hidden_path registry address to be obviously identical to the address configured for the CS... much less confusing.