ci: upload binaries as artifacts
Upload binaries built in the build step as build artifacts.

Inlcudes some cleanup around the uploading of artifacts and related
hooks.

Now only the handling of the bazel-remote and go-module-proxy is left in
the pre-command/pre-exit hooks.
The rest is moved into step-specific metahook-hooks.

The log artifacts are no longer combined into a single tar.gz.
That has never seemed very convenient to me. Let's see if this works well.