ci: fixes for updated buildkite elastic stack version  (#4397)
Fixes for update of buildkite elastic stack to v6.7.1.

Update package "rhel" dependency list for Amazon Linux 2023:
- `bridge-utils`: no longer available. Replace deprecated `brctl` with iproute2's `bridge`
- `moreutils`: no longer available, not used
- `sqlite3`: no longer available, not used
- `ethtool`: required for router_multi test, seems to have previously been installed by default.

For the router_multi test, set the default TTL to 64, so that the packets sent back by the router match the expected value. This is the default value on many other systems, so we did not previously need to worry about this.
The value is only set in a net-namespace, so we don't need to worry about resetting it.