build: made scion.sh mstatus resitant to docker compose cosmetic changes.
The script was parsing the output of docker-compose. This was already
incompatible with the latest docker-compose-v1 (possibly not affecting
the buildkite environment) not mentionning docker compose v2.

I preserved the original mstat behavior, although not all of the intent
is very obvious: why output something only about what is *not* running?