build: Provide a proper home dir to python tests that run "docker compose".
This is due to a bug in some version of docker-compose v2 that causes
an unused .cache directory to be created in $HOME. Due to a bug in
Bazel, tests don't have a HOME variable (it should be set to $TEST_TMPDIR
but isn't set at all). So we must provide one ourselves.

My previous work around of using /tmp had the potential of
creating collisions in the future. This method is cleaner. It uses
an output location internal to the test. This is the simplest I could
manage. Tools are supposed to be enablers, but Bazel prefers to be
a disabler it seems.