router: initial draft of local/transit/incoming/outgoing awareness in metrics.
Packet counters are labeled with a "crossing" flag, which denotes
whether a packet is crossing the local AS boundary or not.

Combined the existing separate ingress/egress counters and the existing
interface labels (from which we can tell internal from external interfaces),
this allows to counts packets according to 4 types of processing:

incoming: from another AS to the local AS
outgoing: from the local AS to another AS
local: from the local AS to the local AS
transit: from another AS to another AS

It should be noted that these counters are about a single router,
so a packet that transits through an AS via two routers will be
counted by each router independently: once as incoming and once as
outgoing. Not as transit.