router: reduce the cost of the processmetrics collector.
The collected metrics are originally per-threads. Collecting them means
knowing what threads exist. The procfs API isn't smart about it and generates
work for the GC. To prevent it, we stat /proc/pid/tasks ourselves and check if
the count of files has changed. Go never kills threads; only adds them. So
any set of changes is a change of count.

Added a metric to get an idea of how often we actually have to pay the price:
process_metrics_tasklist_updates_total