docker compose, drop --compatibility flag
The compatibility flag affects what word separator is used in the
container name.

> Compose generates container names based on the project name, service name, and scale/replica count.
> In Compose V1, an underscore (_) was used as the word separator. In Compose V2, a hyphen (-) is used as the word separator.

We don't usually rely on specific container names, so this should be
fine.
In some exception cases where this seems more practical (containers for
bazel-remote-cache and go-module-proxy) due to special casing in the CI
scripts, container_name is set explicitly.