deps: update oapi-codegen to v2
The oapi-codegen libraries went through a reorganization; the
repositories for the oapi-codegen build tool was separated from the
runtime libraries.
In our go.mod / go_deps.bzl, oapi-codegen build tool no longer shows up.
Instead, we explicitly need to require it, in a new macro rules_openapi_dependencies().

Aside; fix `cfg = "exec"` for the oapi-codegen buildtool so that this
tool is built for the environment of the builder, not the target
platform.