connectrpc: add support for server and client side
Add connectrpc server support for all endpoints with exception
of the hidden path inter AS RPCs.

On the client side, we use happy eyeballs with currently a fixed
delay of 500ms to connect over the legacy gRPC SCION.
This might be tuned in a follow up commit.

The intra AS RPCs are also support connectrpc, but the client
side is not adapted. Here, we anyway have gRPC/TCP at the moment
and moving to connectrpc is less urgent.

Co-authored-by:  Katya Titkova <katyatitkova@gmail.com>
Co-authored-by:  Jean-Christophe Hugly <jice@scion.org>