router: improvements to newbenchmark.
Moved knowledge of mac addresses outside brload. The test harness now tells
brload what the interface mac addresses are and not the other way around.
In the usual automated test context it makes no difference (the same naming
and addressing scheme is used), but in case we ever run the test on a real
self-contained machine, the test doesn't get to choose the interface names
and macs.