router: Make the benchmark and the router a little more adaptive to core count.
In the router. Compute a better default for the number of processors based on multiple
observations: given enough streams to occupy all processors, performance is degraded
if there are as many processors as cores. Best performance is obtained by setting
aside two cores for other tasks. Making this the default simplifies the benchmarking
process by not having to configure an override for each type of target machine being
tested.

In the benchmark. Run the test with many streams. It reflects reality and allows
a very parallel router to perform at its best without distorting the performance
of a less-parallel one. Make sure that the number of streams is likely to be a
multiple of the number of cores so the cores are loaded evenly in spite of
there being typically fewer streams than in real life.
(TBD, learn the router's real config and choose a multiple).