doc: update buildkite instance type (#4459)
Documenting that we've bumped the buildkite instance type from t3 to
m6i.
The main reason for the switch is that the "burstable" nature of the
t*-instances makes them extremely unsuitable for running benchmarks. The
m6i are also much faster, but not much more expensive.
Also, the agent pool is now at most 6 agents. Spinning up 10 agents and
having to warm up all the caches, and then scaling back in after a
single build, seemed excessive.