Work in progress. Create ipk packages.
The method is to import openwrt_SDK as an external dependency and
slap a bazel build file on in.

The rules in that BUILD (and associated .bzl) file encapsulate a
shell script that follows the basic recipe for creating new packages
in the openwrt sdk. (https://digggy.medium.com/openwrt-quick-start-493e08ed73f
is a nice concise one).
The deviation from that recipe is that the binary to be packaged isn't
built within openwrt, it is copied from the scion build.

Building anything in a non-bazel tree while using bazel is difficult:
basel wants you to declare everything you'll touch so it can copy it to
a sandbox....no-can-do. So instead I run the build rule with "no-sandbox".

There are some other hoops to jump through due to cryptic paths shenanigans.
Tried to put comments where it mattered.